Add identical-vern first phase to DupFinder.
