+++ Backend/Interfaces/IMergeService.cs
@@ -10,6 +10,7 @@ namespace BackendFramework.Interfaces
         Task<MergeBlacklistEntry> AddToMergeBlacklist(string projectId, string userId, List<string> wordIds);
         Task<bool> IsInMergeBlacklist(string projectId, List<string> wordIds, string? userId = null);
         Task<int> UpdateMergeBlacklist(string projectId);
-        Task<List<List<Word>>> GetPotentialDuplicates(string projectId, string? userId = null);
+        Task<List<List<Word>>> GetPotentialDuplicates(
+            string projectId, int maxInList, int maxLists, string? userId = null);
     }
 }
